+++ Changes.md
@@ -373,7 +373,11 @@ Build
 - USD : Updated to version 23.05.
 - ZLib : Added version 1.2.13.
 
-1.2.10.x (relative to 1.2.10.2)
+1.2.10.x (relative to 1.2.10.3)
+========
+
+
+1.2.10.3 (relative to 1.2.10.2)
 ========
 
 Fixes